WIP The santa can send a message to their target

+++ database/migrations/2022_01_02_001503_remove_draw_expiration.php
@@ -14,13 +14,11 @@ return new class extends Migration
      */
     public function up()
     {
-        if (! Schema::hasColumn('draws', 'finished_at')) {
+        if (Schema::hasColumn('draws', 'expires_at')) {
             Schema::table('draws', function (Blueprint $table) {
                 $table->date('finished_at')->nullable();
             });
-        }
 
-        if (Schema::hasColumn('draws', 'expires_at')) {
             if(config('app.env') !== 'testing') {
                 DB::table('draws')
                     ->update(
@@ -44,22 +42,27 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::table('draws', function (Blueprint $table) {
-            $table->date('expires_at')->nullable();
-        });
+        if (Schema::hasColumn('draws', 'finished_at')) {
+            Schema::table('draws', function (Blueprint $table) {
+                $table->date('expires_at')->nullable();
+            });
 
-        if(config('app.env') !== 'testing') {
-            DB::table('draws')
-                ->update(
-                    array(
-                        'expires_at' => DB::raw('GREATEST(finished_at, DATE_ADD(updated_at, INTERVAL '.Draw::MONTHS_BEFORE_EXPIRATION.' MONTH))'),
-                    )
-                );
+            if(config('app.env') !== 'testing') {
+                DB::table('draws')
+                    ->update(
+                        array(
+                            'expires_at' => DB::raw('GREATEST(finished_at, DATE_ADD(updated_at, INTERVAL '.Draw::MONTHS_BEFORE_EXPIRATION.' MONTH))'),
+                        )
+                    );
+            }
+
+            Schema::table('draws', function (Blueprint $table) {
+                $table->date('expires_at')->change();
+
+                if (Schema::hasColumn('draws', 'finished_at')) {
+                    $table->dropColumn('finished_at');
+                }
+            });
         }
-
-        Schema::table('draws', function (Blueprint $table) {
-            $table->date('expires_at')->change();
-            $table->dropColumn('finished_at');
-        });
     }
 };
